pyproject.toml
[project]
name = "cs336-systems"
version = "1.0.5"
description = "CS336 Assignment 2 - Systems"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "cs336-basics",
    "humanfriendly",
    "matplotlib",
    "numpy",
    "pytest>=8",
    "python-dotenv",
    "regex>=2024.11.6",
    "torch~=2.6.0; sys_platform != 'darwin' or platform_machine != 'x86_64'",
    "torch~=2.2.2; sys_platform == 'darwin' and platform_machine == 'x86_64'", # Intel Macs
    "tqdm>=4.67.1",
    "wandb>=0.19.7",
    "pandas>=2",
    "ty>=0.0.1a16",
]

[build-system]
requires = ["uv_build>=0.8.3,<0.9.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-name = "cs336_systems"
module-root = ""

[tool.uv]
package = true
python-preference = "managed"

[tool.uv.sources]
cs336-basics = { path = "./cs336-basics", editable = true }  # Change this path to your assignment1-basics repo you want to use your own implementation!



[tool.pytest.ini_options]
log_cli = true
log_cli_level = "WARNING"
addopts = "-s"

[tool.ruff]
line-length = 120

[tool.ruff.lint.extend-per-file-ignores]
# Also ignore `E402` in all `__init__.py` files.
"__init__.py" = ["E402", "F401", "F403", "E501"]

[tool.ruff.lint]
extend-select = ["UP"]
ignore = [
    "F722"
]
